requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.4
numpy>=1.26.2

# Web Scraping & Data Collection
beautifulsoup4>=4.12.2
httpx>=0.25.2
aiohttp>=3.9.1

# Database & Vector Store
chromadb>=0.4.22
sqlite3-api>=2.0.4

# LLM & AI
anthropic>=0.39.0
groq>=0.4.2
langchain>=0.1.0
langchain-anthropic>=0.1.1
langchain-community>=0.0.13

# Web Framework
streamlit>=1.29.0
fastapi>=0.108.0
uvicorn>=0.25.0

# Data Processing
schedule>=1.2.1
pyyaml>=6.0.1

# Testing
pytest>=7.4.3